Services - Usando injeçao de dependencia corretamente, por meio de interfaces (reduzindo acoplamento da implementaçao concreta)
HttpPulsarPay - classe muito grande, refatorei e quebrei em classes menores e mais especificas (boleto, cliente e produto)
