+++ src/main/java/br/com/ericsoncbizarro/IntegracaoPulsarPay/services/HttpPulsarPayBoleto.java
@@ -1,0 +1,10 @@
+package br.com.ericsoncbizarro.IntegracaoPulsarPay.services;
+
+import br.com.ericsoncbizarro.IntegracaoPulsarPay.model.modelPulsarPay.Boleto;
+import java.util.List;
+
+public interface HttpPulsarPayBoleto {
+
+    List<Boleto> getBoletos() throws Exception;
+    List<Boleto> postBoleto(Boleto boletoBody) throws Exception;
+}
