+++ src/main/java/br/com/ericsoncbizarro/IntegracaoPulsarPay/services/HttpPulsarPayCliente.java
@@ -1,0 +1,10 @@
+package br.com.ericsoncbizarro.IntegracaoPulsarPay.services;
+
+import br.com.ericsoncbizarro.IntegracaoPulsarPay.model.modelPulsarPay.Cliente;
+import java.util.List;
+
+public interface HttpPulsarPayCliente {
+
+    List<Cliente> getClientes() throws Exception;
+    List<Cliente> postCliente(Cliente clienteBody) throws Exception;
+}
